Edit add function in app.js
