+++ src/index.html
@@ -1,37 +1,29 @@
 <!DOCTYPE html>
 <html lang="en">
-
-<head>
-  <meta charset="UTF-8" />
-  <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  <script src="https://kit.fontawesome.com/d7d1a0422d.js" crossorigin="anonymous"></script>
-  <title>To Do List App</title>
-<!-- <script defer src="main.js"></script></head> -->
-
-<body>
-  <h2 class="title">My To Do App</h2>
-  <div class="main">
-    <div class="header">
-      <h3 class="tasksTitle">Today's To Do List</h3>
-      <i class="fas fa-sync"></i>
-    </div>
-    <div class="addTask">
-      <form class="form">
-
-        <input id="input" type="text" placeholder="Add to your todo list..." />
-
-        <i class="fas fa-arrow-right" id="submit"></i>
+  <head>
+    <meta charset="UTF-8" />
+    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <link
+      rel="stylesheet"
+      href="https://unicons.iconscout.com/release/v4.0.8/css/line.css"
+    />
+    <title>ToDo-List</title>
+  <body>
+    <div class="container">
+      <div class="title">
+        <h4>Today's ToDo</h4>
+        <i class="uil uil-redo"></i>
+      </div>
+      <hr />
+      <form>
+        <input type="text" id="new-task" placeholder="Add to your list..." />
+        <button type="submit"><i class="uil uil-enter"></i></button>
       </form>
+      <ul id="task-list"></ul>
+
+      <hr />
+      <a href="#">Clear all completed</a>
     </div>
-
-    <div id="myTasksList"></div>
-    <div class="clear">
-      <button class="btn" id="remove">Clear All Completed</button>
-    </div>
-
-
-  </div>
-</body>
-
+  </body>
 </html>
